+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 [Unreleased]
+### Added
+- @babel/polyfill for IE support
+
+### Changed
+- parse S3 XML response on the server side for IE support
+- Use v3.5.2 of vue2-dropzone for IE support
+
 ### Fixed
 - "indexOf" of "null" error if the file type is unknown
 - isset checks for file information which depends on the used browser